Time: 457 ms (55.04%), Space: 151.8 MB (14.22%) - LeetHub

+++ 1930-unique-length-3-palindromic-subsequences/1930-unique-length-3-palindromic-subsequences.cpp
@@ -4,8 +4,8 @@ public:
         int cnt = 0;
         
         for (char c = 'a'; c <= 'z'; ++c) {
-            unsigned long idx1 = s.find(c);
-            unsigned long idx2 = s.rfind(c);
+            unsigned int idx1 = s.find(c);
+            unsigned int idx2 = s.rfind(c);
             
             if (idx1 == string::npos or (idx2 - idx1 + 1) < 3) {
                 continue;
